atualização: Deletando categoria com DELETE

+++ backend/src/main/java/com/jefersonwvs/dscatalog/services/CategoryService.java
@@ -6,10 +6,10 @@ import javax.persistence.EntityNotFoundException;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.dao.DataIntegrityViolationException;
-import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.jefersonwvs.dscatalog.dto.CategoryDTO;
@@ -27,7 +27,7 @@ public class CategoryService {
 	@Transactional(readOnly = true)
 	public Page<CategoryDTO> findAllPaged(Pageable pageable) {
 		Page<Category> list = repository.findAll(pageable);
-		return list.map(x -> new CategoryDTO(x));
+		return list.map(CategoryDTO::new);
 	}
 
 	@Transactional(readOnly = true)
@@ -57,13 +57,15 @@ public class CategoryService {
 		}
 	}
 
+	@Transactional(propagation = Propagation.SUPPORTS)
 	public void delete(Long id) {
+		if (!repository.existsById(id)) {
+			throw new ResourceNotFoundException("Recurso não encontrado");
+		}
 		try {
 			repository.deleteById(id);
-		} catch (EmptyResultDataAccessException e) {
-			throw new ResourceNotFoundException("Id " + id + " não encontrado!");
 		} catch (DataIntegrityViolationException e) {
-			throw new DatabaseException("Violação de integridade!");
+			throw new DatabaseException("Falha de integridade referencial");
 		}
 	}
